requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
greenlet==3.3.2
idna==3.11
playwright==1.58.0
psycopg==3.3.3
psycopg-binary==3.3.3
pyee==13.0.1
python-dotenv==1.2.2
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
